+++ packages/kitchensink/service-fastify-healthz/service-fastify-healthz.test.js
@@ -1,0 +1,6 @@
+describe('service-fastify-healthz', () => {
+  it('Should render the default healthcheck route', async () => {
+    const res = await get(`/healthz`);
+    // expect(res).toContain('index');
+  });
+});