+++ packages/service-fastify-healthz/src/hooks.js
@@ -1,9 +1,9 @@
-const { SERVICE } = require('@forrestjs/hooks');
-
-const SERVICE_NAME = `${SERVICE} fastify-healthz`;
+const SERVICE_NAME = `fastify-healthz`;
 const FASTIFY_HEALTHZ_HANDLER = `${SERVICE_NAME}/hanlder`;
+const FASTIFY_HEALTHZ_CHECK = `${SERVICE_NAME}/check`;
 
 module.exports = {
   SERVICE_NAME,
   FASTIFY_HEALTHZ_HANDLER,
+  FASTIFY_HEALTHZ_CHECK,
 };