+++ packages/kitchensink/CHANGELOG.md
@@ -8,7 +8,8 @@
 - Added examples for `@forrestjs/service-hash`
 - Added examples for `@forrestjs/service-fetchq`
 - Added examples for `@forrestjs/service-logger`
-- Added examples for `@forrestjs/service-fetchq-static`
-- Added examples for `@forrestjs/service-fetchq-cors`
-- Added examples for `@forrestjs/service-fetchq-cookie`
+- Added examples for `@forrestjs/service-fastify-static`
+- Added examples for `@forrestjs/service-fastify-cors`
+- Added examples for `@forrestjs/service-fastify-cookie`
+- Added examples for `@forrestjs/service-fastify-healthz`
 - Added `app01` examples for a complex app